requirements.txt
locustio==0.7.3
pyzmq==15.0.0
websocket-client==0.34.0
